# php-tutor/php-learn/app/cache/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "php-learn API",
        "description": "PHP 8.5 无框架后端 + Vue 3 SPA 前后端分离项目的 REST API。\n\n鉴权方式：Bearer Token（JWT Access + Refresh Token）\n响应结构：统一 envelope {code, message, data, errors?}\n业务 code：0=成功, 40101=未登录, 40102=token 无效, 40301=无权限, 40401=资源不存在, 42201=校验失败, 50000=内部错误",
        "contact": {
            "name": "php-learn",
            "url": "https:\/\/github.com\/example\/php-learn"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "\/api\/v1",
            "description": "API v1（同域部署）"
        },
        {
            "url": "http:\/\/127.0.0.1:8000\/api\/v1",
            "description": "本地开发"
        }
    ],
    "paths": {
...
        "\/auth\/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "登录",
                "description": "用登录名和密码换取 Access + Refresh Token",
                "operationId": "c542ba0f20b129ebb73c48d9561d560b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "登录名（用户名或邮箱）",
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "description": "密码",
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "required": [
                                        "accessToken",
                                        "refreshToken"
                                    ],
                                    "properties": {
                                        "accessToken": {
                                            "description": "Access Token（JWT，15 分钟有效）",
                                            "type": "string"
                                        },
                                        "refreshToken": {
                                            "description": "Refresh Token（7 天有效）",
                                            "type": "string"
                                        },
                                        "tokenType": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expiresIn": {
                                            "description": "Access Token 剩余秒数",
                                            "type": "integer",
                                            "example": 900
                                        },
                                        "user": {
                                            "$ref": "#\/components\/schemas\/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "账号或密码错误",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "required": [
                                        "code",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 40101
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "账号或密码错误"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "required": [
                                        "code",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 42201
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "字段级错误",
                                            "type": "object",
                                            "nullable": true,
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/auth\/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "刷新令牌",
                "description": "用 Refresh Token 换取新的 Access Token（同时轮转 Refresh Token）",
                "operationId": "7f8016872ebe19d2ce467c9feab85394",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "required": [
                                    "refreshToken"
                                ],
                                "properties": {
                                    "refreshToken": {
                                        "description": "刷新令牌",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "刷新成功",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "required": [
                                        "accessToken",
                                        "refreshToken"
                                    ],
                                    "properties": {
                                        "accessToken": {
                                            "description": "新的 Access Token",
                                            "type": "string"
                                        },
                                        "refreshToken": {
                                            "description": "轮转后的新 Refresh Token",
                                            "type": "string"
                                        },
                                        "expiresIn": {
                                            "type": "integer",
                                            "example": 900
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refresh Token 无效或已过期",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "required": [
                                        "code",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 40102
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/auth\/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "登出",
                "description": "撤销当前用户的所有 Refresh Token（Access Token 等自然过期）",
                "operationId": "e401b18266e30a630a8c26f097388086",
                "responses": {
                    "204": {
                        "description": "登出成功（无 body）"
                    },
                    "401": {
                        "description": "未登录",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "required": [
                                        "code",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 40101
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "\/auth\/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "获取当前用户",
                "description": "返回当前登录用户的信息（需 Bearer Token）",
                "operationId": "f7ae1fe2dcffa6eafa62de88e6441088",
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或 token 失效",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token（JWT Access Token），通过登录接口获取",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}